# apps/app/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: SkyPlanner API
  description: |
    REST API for SkyPlanner kundeadministrasjon og ruteplanlegging.

    ## Autentisering

    API-et stotter to autentiseringsmetoder:

    ### API-nokkel
    For tredjeparts-integrasjoner. Opprett API-nokkel i admin-panelet.
    ```
    X-API-Key: sk_live_xxx
    ```

    ### JWT Bearer Token
    For innloggede brukere via web-applikasjonen.
    ```
    Authorization: Bearer <token>
    ```

    ## Rate Limiting
    - Standard: 100 requests per 15 minutter
    - API-nokler kan ha egne kvote-innstillinger

    ## Feilhandtering
    Alle feil returnerer et standardisert format:
    ```json
    {
      "success": false,
      "error": {
        "code": "ERROR_CODE",
        "message": "Beskrivende feilmelding"
      },
      "requestId": "abc123"
    }
    ```

  version: 1.0.0
  contact:
    name: Efffekt AS
    email: support@skyplanner.no
    url: https://skyplanner.no
servers:
  - url: https://app.skyplanner.no/api
    description: Produksjonsserver
  - url: http://localhost:3000/api
    description: Lokal utvikling
...
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /v1/customers:
    get:
      tags:
        - Kunder
      summary: List kunder
      description: Henter en paginert liste over kunder
      operationId: listCustomers
      security:
        - ApiKeyAuth: [customers:read]
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/offsetParam'
        - name: search
          in: query
          description: Sok i navn, adresse, telefon, epost
          schema:
            type: string
        - name: kategori
          in: query
          description: Filtrer på kategori
          schema:
            type: string
      responses:
        '200':
          description: Liste over kunder
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          data:
                            type: array
                            items:
                              $ref: '#/components/schemas/Customer'
                          total:
                            type: integer
                          limit:
                            type: integer
                          offset:
                            type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags:
        - Kunder
      summary: Opprett kunde
      description: Oppretter en ny kunde
      operationId: createCustomer
      security:
        - ApiKeyAuth: [customers:write]
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCustomerRequest'
      responses:
        '201':
          description: Kunde opprettet
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Customer'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/customers/{id}:
    get:
      tags:
        - Kunder
      summary: Hent kunde
      description: Henter en spesifikk kunde
      operationId: getCustomer
      security:
        - ApiKeyAuth: [customers:read]
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/idParam'
      responses:
        '200':
          description: Kundedetaljer
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Customer'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Kunder
      summary: Oppdater kunde
      description: Oppdaterer en eksisterende kunde
      operationId: updateCustomer
      security:
        - ApiKeyAuth: [customers:write]
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/idParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCustomerRequest'
      responses:
        '200':
          description: Kunde oppdatert
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Customer'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Kunder
      summary: Slett kunde
      description: Sletter en kunde permanent
      operationId: deleteCustomer
      security:
        - ApiKeyAuth: [customers:write]
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/idParam'
      responses:
        '200':
          description: Kunde slettet
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          message:
                            type: string
                            example: Kunde slettet
        '404':
          $ref: '#/components/responses/NotFound'

  # ============ Webhooks ============
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API-nokkel opprettet i admin-panelet

    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT-token fra innlogging
  parameters:
    idParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
      description: Unik ID
    limitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 100
        maximum: 500
      description: Maks antall resultater
    offsetParam:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
      description: Antall resultater å hoppe over
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        requestId:
          type: string
          example: req_abc123
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Ugyldig data
        requestId:
          type: string

    # ============ Customer Types ============
    Customer:
      type: object
      properties:
        id:
          type: integer
          example: 1
        navn:
          type: string
          example: Ola Nordmann AS
        adresse:
          type: string
          example: Storgata 1
        postnummer:
          type: string
          example: "0123"
        poststed:
          type: string
          example: Oslo
        telefon:
          type: string
          example: "12345678"
        epost:
          type: string
          format: email
          example: post@example.com
        lat:
          type: number
          format: double
          example: 59.9139
        lng:
          type: number
          format: double
          example: 10.7522
        kategori:
          type: string
          example: El-Kontroll
        kontaktperson:
          type: string
        notater:
          type: string
        siste_el_kontroll:
          type: string
          format: date
        neste_el_kontroll:
          type: string
          format: date
        el_kontroll_intervall:
          type: integer
          description: Måneder mellom kontroller
          example: 36
        siste_brann_kontroll:
          type: string
          format: date
        neste_brann_kontroll:
          type: string
          format: date
        brann_kontroll_intervall:
          type: integer
          example: 12
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateCustomerRequest:
      type: object
      required:
        - navn
        - adresse
      properties:
        navn:
          type: string
          minLength: 1
          maxLength: 255
        adresse:
          type: string
        postnummer:
          type: string
          pattern: '^\d{4}$'
        poststed:
          type: string
        telefon:
          type: string
        epost:
          type: string
          format: email
        lat:
          type: number
        lng:
          type: number
        kategori:
          type: string
          enum: [El-Kontroll, Brannvarsling, Begge]
        kontaktperson:
          type: string
        notater:
          type: string
    UpdateCustomerRequest:
      type: object
      properties:
        navn:
          type: string
        adresse:
          type: string
        postnummer:
          type: string
        poststed:
          type: string
        telefon:
          type: string
        epost:
          type: string
          format: email
        kontaktperson:
          type: string
        notater:
          type: string

    # ============ Webhook Types ============
...
  responses:
    BadRequest:
      description: Ugyldig forespørsel
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: "Ugyldig data: navn er påkrevd"
    Unauthorized:
      description: Manglende eller ugyldig autentisering
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: Ugyldig API-nokkel
    Forbidden:
      description: Manglende tilgang
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: FORBIDDEN
              message: Manglende scope for denne operasjonen
    NotFound:
      description: Ressurs ikke funnet
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: NOT_FOUND
              message: Kunde ikke funnet

# ============ Webhook Payload Documentation ============
...